Fix UI issues
