+++ Assets/Prefabs/UI/Menus/Lose Menu.prefab
@@ -209,7 +209,32 @@ PrefabInstance:
     - target: {fileID: 1910789372790776597, guid: 884ab93e68c284946b4f41fd334d713d,
         type: 3}
       propertyPath: m_Name
-      value: Lose Menu
+      value: Death Menu
+      objectReference: {fileID: 0}
+    - target: {fileID: 1910789372790776597, guid: 884ab93e68c284946b4f41fd334d713d,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1910789373094401309, guid: 884ab93e68c284946b4f41fd334d713d,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
+      value: 
+      objectReference: {fileID: 0}
+    - target: {fileID: 1910789373094401309, guid: 884ab93e68c284946b4f41fd334d713d,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
+      value: Menus.PauseMenu, Scripts
+      objectReference: {fileID: 0}
+    - target: {fileID: 2539904799060679499, guid: 884ab93e68c284946b4f41fd334d713d,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
+      value: 
+      objectReference: {fileID: 0}
+    - target: {fileID: 2539904799060679499, guid: 884ab93e68c284946b4f41fd334d713d,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
+      value: Menus.PauseMenu, Scripts
       objectReference: {fileID: 0}
     - target: {fileID: 3276857934266911518, guid: 884ab93e68c284946b4f41fd334d713d,
         type: 3}
@@ -221,11 +246,31 @@ PrefabInstance:
       propertyPath: freezeOnToggle
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 3276857934266911518, guid: 884ab93e68c284946b4f41fd334d713d, type: 3}
+    - {fileID: 1910789372790776598, guid: 884ab93e68c284946b4f41fd334d713d, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 884ab93e68c284946b4f41fd334d713d, type: 3}
+--- !u!1 &5172193196748381800 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 1910789372790776597, guid: 884ab93e68c284946b4f41fd334d713d,
+    type: 3}
+  m_PrefabInstance: {fileID: 6720266198743268733}
+  m_PrefabAsset: {fileID: 0}
 --- !u!224 &5172193196748381807 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 1910789372790776594, guid: 884ab93e68c284946b4f41fd334d713d,
     type: 3}
   m_PrefabInstance: {fileID: 6720266198743268733}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &1812395219
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5172193196748381800}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 62a14d240f444b96a6e9f6eafee22387, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
