+++ Assets/Prefabs/UI/Menus/Options Menu.prefab
@@ -509,8 +509,8 @@ RectTransform:
   m_Father: {fileID: 1779187630408558378}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMin: {x: 1, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0.000009536743}
   m_SizeDelta: {x: 20, y: -29.9219}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -1004,14 +1004,13 @@ GameObject:
   - component: {fileID: 1779187630194480876}
   - component: {fileID: 8605523634544552563}
   - component: {fileID: 8605523634544552562}
-  - component: {fileID: 6153217522618498051}
   m_Layer: 5
   m_Name: Options Menu
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1779187630194480875
 RectTransform:
   m_ObjectHideFlags: 0
@@ -1090,21 +1089,6 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6153217522618498051
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1779187630194480874}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 462cf0baac97aee46a7ec45a3ef7a2fe, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  menuName: options
-  freezeOnToggle: 0
-  inputToggle: 0
 --- !u!1 &1779187630251189476
 GameObject:
   m_ObjectHideFlags: 0
@@ -1355,7 +1339,7 @@ MonoBehaviour:
         m_Mode: 2
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: Menus.Menu, Scripts
+          m_ObjectArgumentAssemblyTypeName: Menus.MenuController, Scripts
           m_IntArgument: 0
           m_FloatArgument: 0
           m_StringArgument: 
@@ -3293,7 +3277,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
