--- Assets/Scripts/Menus/CreateRoomMenu.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 4fcffeb7a97542f29915ec7b8a2c4e69
-timeCreated: 1643796068
