--- Assets/Scripts/Menus/Menu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 462cf0baac97aee46a7ec45a3ef7a2fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
