--- Assets/Scripts/UI/SettingsMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c448b6bc6de81264a9efa9b5f4b16db2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
